589 behavior tree cleanup code (#600)
* behavior tree MD small word fixes

* speed for leaving parking place set from 30kmh to 20

* folder behaviours to behaviors

* enum not needed

* road features to all own classes

* unstuckroutine, cruise and leave parking space separated

* Revert "speed for leaving parking place set from 30kmh to 20"

This reverts commit 5da562b75b9c0eba0cb7e92c2d0e76876cd28cb5.

* delete old  not neededfiles

* docu filepaths inside updated and small correction

---------
